backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
torch==2.6.0
torchvision==0.21.0
Pillow==10.4.0
numpy==1.26.4
opencv-python==4.10.0.84
joblib==1.4.2
scikit-learn==1.5.2
pydantic==2.9.0
python-dotenv==1.0.1